requirements.txt
fastapi
uvicorn
google-genai
itsdangerous===2.2.0
jinja2
python-multipart